fix: apply all review feedback from PR #14
- payment_service: fix get_or_create_wallet savepoint pattern (IntegrityError must be caught outside begin_nested)
- payment_service: check locked_wallet.is_frozen after FOR UPDATE in credit_wallet and debit_wallet
- payment_service: enforce WALLET_NEGATIVE_FLOOR in code before writing to DB in debit_wallet
- payment_service: add amount > 0 validation in create_withdrawal_request
- payment_service: use (description or '') in complete_wallet_topup to avoid None string
- booking_service: wrap settlement in nested transaction savepoint
- provider_wallet: change role code from 'provider' to 'provider_owner'
- internal_payment: update docstring (VNPay callback is GET not POST)
- internal_payment: catch IntegrityError before generic Exception for idempotent VNPay/MoMo/ZaloPay callbacks
- vnpay_service: use Decimal for amount*100 calculation to avoid float rounding
- vnpay_service: use hmac.compare_digest() for timing-safe hash comparison
- schemas/payment: use Decimal instead of float for WalletResponse.balance
- sql: move migrate_006_payment_hardening.sql to established sql/ directory
- sql: remove redundant ix_promotion_usages_user_promo from migrate_006 (already in migrate_005)
- tests: use 'provider_owner' role code in make_provider_token and seed_provider_pair
- tests: create real Booking row in test_consume_promotion_success to satisfy FK constraint

Agent-Logs-Url: https://github.com/thangnqdev/sandichvu/sessions/28d3b352-0391-4993-b2e1-2d5bfd06a6cb

+++ backend/app/schemas/payment.py
@@ -2,6 +2,7 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
+from decimal import Decimal
 from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator
@@ -18,7 +19,7 @@ class WalletResponse(BaseModel):
 
     id: uuid.UUID
     user_id: uuid.UUID
-    balance: float = Field(description="Số dư hiện tại")
+    balance: Decimal = Field(description="Số dư hiện tại")
     currency: str = Field(description="Đơn vị tiền tệ (VND/USD)")
     is_frozen: bool = Field(description="Ví bị đóng băng")
     created_at: datetime
